ENH: Creation of multiple ui files and insertion of new ui files in a QTabWidget.
Each ui file also has a derived class that is a QWidget to be able to call the functions 'show' and 'hide'. This could also be done by using only the ui files and have a main frame in each of them. We could then just show or hide the frame, which would avoid the burden of creating many simple files.